Refactor name and use builder pattern for client
